Fix bug 963097 - Minimize pip-installed packages
Don't install a bunch of other stuff, that might hide problems
during development. If the code depends on some package that's
not in vendor, vendor-local, or requirements/compiled.txt, we
want to find out about it sooner rather than later.

It turned out that nothing in requirements/prod.txt was needed
at all - everything was already in vendor or vendor-local.